TRUNK-4208: Removed gender as a requirement during PersonValidation, added it instead to PatientValidation.
- Moved unit tests from PersonValidatorTest to PatientValidatorTest to reflect this new behaviour.

- Birthdate is already *not* a requirement for Persons.

- Added test for PersonValidator to check that it's not validating gender requirement.

TRUNK-4208: Refactored PersonValidator. Extracted birth date validation into it's own method.
