+++ api/src/test/java/org/openmrs/validator/PatientValidatorTest.java
@@ -168,4 +168,19 @@ public class PatientValidatorTest extends PersonValidatorTest {
 		
 		Assert.assertFalse(errors.hasErrors());
 	}
+	
+	/**
+	 * @see {@link org.openmrs.validator.PatientValidator#validate(Object,Errors)}
+	 * @verifies fail validation if gender is blank
+	 */
+	@Test
+	@Verifies(value = "should fail validation if gender is blank", method = "validate(Object,Errors)")
+	public void validate_shouldFailValidationIfGenderIsBlank() throws Exception {
+		Patient pa = new Patient(1);
+		Errors errors = new BindException(pa, "patient");
+		validator.validate(pa, errors);
+		
+		Assert.assertTrue(errors.hasFieldErrors("gender"));
+	}
+	
 }
